% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_joint.R
\name{simulate_joint_}
\alias{simulate_joint_}
\title{simulate_joint_}
\usage{
simulate_joint_(
  p_control = 0.5,
  p_treatment = 0.65,
  prob_accept_ha = 0.95,
  total_sample_size = 20,
  no_of_sim = 100,
  alternative = "greater",
  prior_p_control = beta_prior(1, 1),
  prior_log_odds_ratio = normal_prior(0, 1),
  n_iter = 1000,
  n_burnin = 30,
  n_thin = 20
)
}
\arguments{
\item{p_control}{}

\item{p_treatment}{}

\item{prob_accept_ha}{}

\item{total_sample_size}{}

\item{no_of_sim}{}

\item{alternative}{}

\item{prior_p_control}{}

\item{prior_log_odds_ratio}{}

\item{n_iter}{}

\item{n_burnin}{}

\item{n_thin}{}
}
\description{
simulate_joint_
}
\examples{

list(p_control = 0.5,
     p_treatment = 0.65,
     prob_accept_ha = 0.95,
     total_sample_size = 20,
     no_of_sim = 100,
     alternative = "greater",
     prior_p_control = beta(1,1),
     prior_log_odds_ratio = normal(0,1)) |> 
  simulate_joint(no_of_sim = 1000)

}
